chore: oss ent restructuring
